+++ Webapp/pages/index/index.js
@@ -1,6 +1,8 @@
 // home.js
-var wxCharts = require('../../utils/wxcharts-min.js');
+var wxCharts = require('../../style/js/wxcharts-min.js');
 var app = getApp();
+//引用promise
+var promise = require("../../utils/wxPromisify.js");
 //调用登录接口
 app.getUserInfo();
 var areaChart = null;
@@ -27,86 +29,105 @@ Page({
     },    
     onLoad: function (e) {
         var self = this;
-
-        //初始化信息，发送请求
-        var http = require("../../utils/httpUtil.js");
-        var params = {
-            id: 1 
-        };
-        var api = "WechatCustomer/";
-        http.GET(api, params, function(res){
-            //如果未成功请求微信服务器
-            if (res.data == "error") {
-                wx.showToast({
-                    title: '请求失败',
-                    icon: 'loading',
-                    duration: 2000
-                })
-                wx.switchTab({
-                  url: '/pages/login/index'
-                })
-            }
-            //判断用户是否已绑定净水器
-            if (res.data == false) {
-                //跳转至登录界面
-                wx.switchTab({
-                  url: '/pages/login/index'
-                })
-            } else {
-                app.info = res.data;
-                console.log(app.info)
-                //更新视图
-                self.setData({
-                    todayUsedWater: app.info.todayUsedWater,
-                    lastUsedWater: app.info.lastUsedWater,
-                    lastFilterChip: app.info.lastFilterChip,
-                    usedBeforeWaterQuality: app.info.usedBeforeWaterQuality,
-                    usedAfterWaterQuality: app.info.usedAfterWaterQuality
-                })
-            }
-        });
         
-        var windowWidth = 320;
-        try {
-          var res = wx.getSystemInfoSync();
-          windowWidth = res.windowWidth;
-        } catch (e) {
-          console.error('getSystemInfoSync failed!');
-        }
-        
-        console.log(app.info);
-        //折线图
-        areaChart = new wxCharts({
-            canvasId: 'areaCanvas',
-            type: 'area',
-            categories: ['02-20', '02-21', '02-22', '02-23', '02-24', '02-25'],
-            animation: true,
-            series: [{
-                name: '用水量(升)',
-                data: [32, 45, 45, 56, 33, 34],
-                format: function (val) {
-                    return val.toFixed(2) + '';
-                }
-            }],
-            yAxis: {
-                title: '',
-                format: function (val) {
-                    return val.toFixed(2);
+        //更新首页数据        
+        initialize().then(function (value) {
+            //获取屏幕宽度
+            try {
+              var res = wx.getSystemInfoSync();
+              var windowWidth = res.windowWidth;
+            } catch (e) {
+              console.error('getSystemInfoSync failed!');
+            }
+            //对获取的日期格式做修改，如2017-05-03改为05-03
+            var keys = [];
+            Object.keys(value).forEach(function (key) {
+                keys.push(key.substring(5, 10));
+            });
+            //获取每天的用水量
+            var datas = [];
+            Object.keys(value).forEach(function (key) {
+                datas.push(value[key]);
+            });
+            //折线图
+            areaChart = new wxCharts({
+                canvasId: 'areaCanvas',
+                type: 'area',
+                categories: keys,
+                animation: true,
+                series: [{
+                    name: '用水量(升)',
+                    data: datas,
+                    format: function (val) {
+                        return val.toFixed(2) + '';
+                    }
+                }],
+                yAxis: {
+                    title: '',
+                    format: function (val) {
+                        return val.toFixed(2);
+                    },
+                    min: 0,
+                    fontColor: '#8085e9',
+                    gridColor: '#8085e9',
+                    titleFontColor: '#f7a35c'
                 },
-                min: 0,
-                fontColor: '#8085e9',
-                gridColor: '#8085e9',
-                titleFontColor: '#f7a35c'
-            },
-            xAxis: {
-                fontColor: '#7cb5ec',
-                gridColor: '#7cb5ec'
-            },
-            extra: {
-                legendTextColor: '#cb2431'
-            },
-            width: windowWidth,
-            height: 200
+                xAxis: {
+                    fontColor: '#7cb5ec',
+                    gridColor: '#7cb5ec'
+                },
+                extra: {
+                    legendTextColor: '#cb2431'
+                },
+                width: windowWidth,
+                height: 200
+            });
+        }).catch(function (error) {
+            console.log(error);
         });
+
+        //获取最近一周用水量
+        function initialize() {
+            return new Promise(function (resolve, reject) {
+                //初始化信息，发送请求
+                var http = require("../../utils/httpUtil.js");
+                var params = {
+                    id: 1 
+                };
+                var api = "WechatCustomer/";
+                http.GET(api, params, function(res){
+                    //如果未成功请求微信服务器
+                    if (res.data == "error") {
+                        wx.showToast({
+                            title: '请求失败',
+                            icon: 'loading',
+                            duration: 2000
+                        })
+                        wx.switchTab({
+                          url: '/pages/login/index'
+                        })
+                    }
+                    //判断用户是否已绑定净水器
+                    if (res.data == false) {
+                        //跳转至登录界面
+                        wx.switchTab({
+                          url: '/pages/login/index'
+                        })
+                    } else {
+                        app.info = res.data;
+                        //更新视图
+                        self.setData({
+                            todayUsedWater: app.info.todayUsedWater,
+                            lastUsedWater: app.info.lastUsedWater,
+                            lastFilterChip: app.info.lastFilterChip,
+                            usedBeforeWaterQuality: app.info.usedBeforeWaterQuality,
+                            usedAfterWaterQuality: app.info.usedAfterWaterQuality
+                        })
+                    }
+                    resolve(res.data.recentOneWeekUsedWater);
+                });
+            });
+        }
     }
 });
+
