+++ Webapp/utils/util.js
@@ -19,29 +19,3 @@ function formatNumber(n) {
 module.exports = {
   formatTime: formatTime
 }
-
-//使用promise
-var Promise = require('bluebird.min')
-
-function wxPromisify(fn) {  
-  return function (obj = {}) {    
-    return new Promise((resolve, reject) => {      
-      obj.success = function (res) {        
-        resolve(res)      
-      }      
-
-      obj.fail = function (res) {        
-        reject(res)      
-      }      
-
-      fn(obj)    
-    })  
-  }
-}
-
-module.exports = {  
-  wxPromisify: wxPromisify
-}
-// var util = require('../../utils/util')
-// var t = util.wxPromisify(app.getUserInfo);
-// t().then(test1());
